Solved: Sqrt(x) (#69)

+++ problems/0069-sqrtx/solution.cpp
@@ -4,9 +4,9 @@ public:
         int low = 0, high = x;
         while(low <= high){
             int mid = (low + high)/2;
-            if((mid*mid) == x) return mid;
-            else if(((mid*mid) < x) && ((mid + 1)*(mid + 1) > x)) return mid;
-            else if((mid*mid) < x) low = mid + 1;
+            if((long long)mid*mid == x) return mid;
+            else if(((long long)mid*mid < x) && ((long long)(mid + 1)*(mid + 1) > x)) return mid;
+            else if((long long)mid*mid < x) low = mid + 1;
             else high = mid - 1;
         }
         return -1;
